Mejoras en la gestión de recursos y lógica del juego
Se introdujeron las clases `ResourceManager` y `AssetsManager` para optimizar la carga y gestión de texturas y sonidos. Se mejoró la lógica de enemigos y jugador, así como la gestión de niveles y power-ups. Se añadieron funciones para crear activos de marcador de posición y se optimizó el código existente para mejorar la legibilidad y el rendimiento. Se corrigieron errores y se mejoró la estructura general del código para facilitar su mantenimiento y expansión futura.

+++ src/intro.cpp
@@ -21,8 +21,8 @@ void FadeInOutLogo(Texture2D texture, const char* companyText, int fadeDuration,
         logoScale = static_cast<float>(maxLogoHeight) / texture.height;
     }
 
-    int logoDrawWidth = texture.width * logoScale;
-    int logoDrawHeight = texture.height * logoScale;
+    int logoDrawWidth = static_cast<int>(texture.width * logoScale);
+    int logoDrawHeight = static_cast<int>(texture.height * logoScale);
 
     int headerContentHeight = logoDrawHeight + marginBetweenLogoAndText + companyTextSize;
     int headerContentY = (screenHeight - headerContentHeight) / 2;
@@ -160,9 +160,9 @@ void ShowIntro(int screenWidth, int screenHeight, Sound fxStart)
     SetSoundVolume(fxStart, 20); // ESTA LINEA PARA PODER SUBIR EL VOLUMEN DEL SONIDITO START
     PlaySound(fxStart);
 
-    // Parpadeo rápido de "press_start" al presionar espacio
+    // Parpadeo rï¿½pido de "press_start" al presionar espacio
 
-    for (int i = 0; i < 10; i++) // Aumentar el número de iteraciones para que dure más tiempo
+    for (int i = 0; i < 10; i++) // Aumentar el nï¿½mero de iteraciones para que dure mï¿½s tiempo
     {
         BeginDrawing();
         ClearBackground(BLACK);
